+++ benchmark/src/androidTest/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.github.sckm.itemgroup.example_benchmark">
+    package="com.github.sckm.itemgroup.benchmark">
 
     <!-- Important: disable debuggable for accurate performance results -->
     <application
