+++ benchmark/src/androidTest/java/com/github/sckm/itemgroup/benchmark/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.example_benchmark
+package com.github.sckm.itemgroup.benchmark
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.example_benchmark.test", appContext.packageName)
+        assertEquals("com.github.sckm.itemgroup.benchmark.test", appContext.packageName)
     }
 }
