+++ benchmark/src/main/res/values/strings.xml
