+++ benchmark/src/test/java/com/github/sckm/itemgroup/benchmark/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.example_benchmark
+package com.github.sckm.itemgroup.benchmark
 
 import org.junit.Test
